backend/requirements.txt
# Shared requirements for all services
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.0
python-dotenv>=1.0.0
httpx>=0.25.0
python-socketio>=5.10.0
pyotp>=2.9.0
qrcode[pil]>=7.4.0
cryptography>=41.0.0
google-generativeai>=0.3.0
pydantic[email]>=2.5.0
